+++ src/hint_processor/builtin_hint_processor/hint_code.rs
@@ -283,6 +283,9 @@ ids.carry_low = 1 if sum_low >= ids.SHIFT else 0
 sum_high = ids.a.high + ids.b.high + ids.carry_low
 ids.carry_high = 1 if sum_high >= ids.SHIFT else 0"#;
 
+pub const UINT128_ADD: &str = r#"res = ids.a + ids.b
+ids.carry = 1 if res >= ids.SHIFT else 0"#;
+
 pub const UINT256_SUB: &str = r#"def split(num: int, num_bits_shift: int = 128, length: int = 2):
     a = []
     for _ in range(length):
@@ -307,6 +310,12 @@ root = isqrt(n)
 assert 0 <= root < 2 ** 128
 ids.root.low = root
 ids.root.high = 0"#;
+
+pub const UINT256_SQRT_FELT: &str = r#"from starkware.python.math_utils import isqrt
+n = (ids.n.high << 128) + ids.n.low
+root = isqrt(n)
+assert 0 <= root < 2 ** 128
+ids.root = root"#;
 
 pub const UINT256_SIGNED_NN: &str = "memory[ap] = 1 if 0 <= (ids.a.high % PRIME) < 2 ** 127 else 0";
 
@@ -850,6 +859,49 @@ ids.remainder.d1 = remainder_split[1]
 ids.remainder.d2 = remainder_split[2]";
 pub const UINT384_SIGNED_NN: &str = "memory[ap] = 1 if 0 <= (ids.a.d2 % PRIME) < 2 ** 127 else 0";
 
+pub(crate) const GET_SQUARE_ROOT: &str =
+    "from starkware.python.math_utils import is_quad_residue, sqrt
+
+def split(num: int, num_bits_shift: int = 128, length: int = 3):
+    a = []
+    for _ in range(length):
+        a.append( num & ((1 << num_bits_shift) - 1) )
+        num = num >> num_bits_shift
+    return tuple(a)
+
+def pack(z, num_bits_shift: int = 128) -> int:
+    limbs = (z.d0, z.d1, z.d2)
+    return sum(limb << (num_bits_shift * i) for i, limb in enumerate(limbs))
+
+
+generator = pack(ids.generator)
+x = pack(ids.x)
+p = pack(ids.p)
+
+success_x = is_quad_residue(x, p)
+root_x = sqrt(x, p) if success_x else None
+
+success_gx = is_quad_residue(generator*x, p)
+root_gx = sqrt(generator*x, p) if success_gx else None
+
+# Check that one is 0 and the other is 1
+if x != 0:
+    assert success_x + success_gx ==1
+
+# `None` means that no root was found, but we need to transform these into a felt no matter what
+if root_x == None:
+    root_x = 0
+if root_gx == None:
+    root_gx = 0
+ids.success_x = int(success_x)
+split_root_x = split(root_x)
+split_root_gx = split(root_gx)
+ids.sqrt_x.d0 = split_root_x[0]
+ids.sqrt_x.d1 = split_root_x[1]
+ids.sqrt_x.d2 = split_root_x[2]
+ids.sqrt_gx.d0 = split_root_gx[0]
+ids.sqrt_gx.d1 = split_root_gx[1]
+ids.sqrt_gx.d2 = split_root_gx[2]";
 pub const HI_MAX_BITLEN: &str =
     "ids.len_hi = max(ids.scalar_u.d2.bit_length(), ids.scalar_v.d2.bit_length())-1";