+++ src/hint_processor/builtin_hint_processor/uint384.rs
@@ -196,9 +196,9 @@ pub fn add_no_uint384_check(
 
     let sum_d0 = a.d0.to_biguint() + b.d0.to_biguint();
     let carry_d0 = Felt252::from((sum_d0 >= shift) as usize);
-    let sum_d1 = a.d1.to_biguint() + b.d1.to_biguint();
+    let sum_d1 = a.d1.to_biguint() + b.d1.to_biguint() + carry_d0.to_biguint();
     let carry_d1 = Felt252::from((sum_d1 >= shift) as usize);
-    let sum_d2 = a.d2.to_biguint() + b.d2.to_biguint();
+    let sum_d2 = a.d2.to_biguint() + b.d2.to_biguint() + carry_d1.to_biguint();
     let carry_d2 = Felt252::from((sum_d2 >= shift) as usize);
 
     insert_value_from_var_name("carry_d0", carry_d0, vm, ids_data, ap_tracking)?;